Bugfix #2
Fixed bug where the random walk could visit unphysical area of the
parameter space. The loglikelihood now tests for flatness of the
spectrum, which is not wanted. This is done in a very crude way, by
testing the difference between the maximum and minimum in the area to be
evaluated. The attribute `atol` controls the strength of the absolute
allowed difference. The testing is done in each array returned by the
method `seperateResponse`, meaning IsomerSpectrum and CombinedSpectrum
should also benefit from this test.